+++ Assets/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bbddc100d27bdf440adcf4d9ff856825
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
