+++ Assets/New Terrain 1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3e13cf277f3e76c4aa2ac8652700f87b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
