+++ Assets/Scripts/Sheep.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c40974f6470d6414194cca8087222682
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
