+++ Assets/Scripts/Sheep/SheepNavigationTest.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.AI;
+
+public class SheepNavigationTest : MonoBehaviour
+{
+    public Transform goal;
+
+    void Start()
+    {
+        NavMeshAgent agent = GetComponent<NavMeshAgent>();
+        agent.destination = goal.position;
+    }
+}
