--- Assets/SheepNavigationTest.cs
@@ -1,15 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.AI;
-
-public class SheepNavigationTest : MonoBehaviour
-{
-    public Transform goal;
-
-    void Start()
-    {
-        NavMeshAgent agent = GetComponent<NavMeshAgent>();
-        agent.destination = goal.position;
-    }
-}
