--- Assets/SheepNavigationTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9c05de199ede1b6408f2247213ad85fc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
